Fixed client crash when message to send is bigger than the shared memory size
Sometimes if the server crashed and the client was still pumping traces,
it would accumulate to a size that would be bigger than the shared memory
causing a crash. I have added code to simply return and clear the traces
if the message is to big. We might want to chunk it up into smaller
chunks and send them separately