Fix race condition when initializing the shared memory
The race condition happened because we were directly getting the pointer
from the shared memory and keeping them as member variable. Thus another
thread could request the pointer from the shared memory and assume it
was initialized because the pointer was not NULL, even though the
ring buffer was not yet initialized. To fix this, we simply expose
the poiter once we have initialized everything