Fixed window build
In windows the shared memory name and the shared semaphore name cannot
be the same. This is what was making hop crashed. I have added simple
code to create a unique name for the semaphore. I have also added code
to handle a failed initialization of HOP so the client does not crash.
(He will instead get spammed by console error msg...)