Fixed issue where set_thread_name would create worker
When calling set_thread_name on a new thread, a worker would
be created thus reducing the amount of worker (thread) available.

We now only create the worker if other data is sent by this thread.
We save the name and set it as the current name when we create the
worker