Added smaller handle for track resize (#23)
When there is a lot of shallow threads, it can be difficult to navigate
without accidently resizing tracks. I have created a handle on the far
right of each tracks that can be used to resize the track and remove
the ability to resize the track from anywhere on the separator.