Added clearer message when no lock owner is found
What I thought was a bug is actually intended behavior. No lock
owner might appear when the actual acquire of the lock is smaller
than the value specified by HOP_MIN_LOCK_CYCLES. This might show
up as a large trace in hop since the os might not immediately
give the lock to the other thread.