Removed shared semaphore
It was brought to my attention (Thanks Pierre), that the shared semaphore
is probably not needed in a lockless context. Although I don't think it
affects the 'lockless' nature of the ringbuffer (since it only signals
the semaphore and never waits on it), it is simply not needed. At first
I thought it was needed for memory coherency between the process, but
it seems the OS is already handling that like a champ !